refactor: TestBuilder 중복 코드 리팩토링 (#10)
