+++ lib/dsl/test-builders/RequestBuilder.ts
@@ -14,25 +14,16 @@
  * limitations under the License.
  */
 
-import { HttpMethod } from "../enums"
-import { PATH_PARAM_TYPES, QUERY_PARAM_TYPES, TestCaseConfig } from "./TestCaseConfig"
+import { PATH_PARAM_TYPES, QUERY_PARAM_TYPES } from "./TestCaseConfig"
 import { DSLField } from "../interface"
 import { ResponseBuilder } from "./ResponseBuilder"
 import { FIELD_TYPES } from "../interface/field"
+import { AbstractTestBuilder } from "./AbstractTestBuilder"
 
-export class RequestBuilder {
-    private readonly config: TestCaseConfig
-    private readonly method: HttpMethod
-    private readonly url: string
-    private readonly app: any
-
-    public constructor(defaults: TestCaseConfig = {}, method: HttpMethod, url: string, app: any) {
-        this.config = { ...defaults }
-        this.method = method
-        this.url = url
-        this.app = app
-    }
-
+/**
+ * API 요청 정보를 설정하는 빌더 클래스입니다.
+ */
+export class RequestBuilder extends AbstractTestBuilder {
     /**
      * 요청시 사용할 헤더를 설정합니다.
      * @param headers
