+++ lib/dsl/test-builders/TestCaseConfig.ts
@@ -14,12 +14,6 @@
  * limitations under the License.
  */
 
-/**
- * API 테스트를 위한 헬퍼 클래스와 타입 정의
- * @description
- * API 테스트를 쉽게 구성하고 실행할 수 있도록 도와주는 빌더 패턴 기반의 헬퍼 클래스입니다.
- */
-
 import { HttpStatus } from "../enums"
 import { DSLField } from "../interface"
 import { FIELD_TYPES } from "../interface/field"
@@ -27,6 +21,9 @@ import { FIELD_TYPES } from "../interface/field"
 export type PATH_PARAM_TYPES = string | number
 export type QUERY_PARAM_TYPES = string | number | boolean
 
+/**
+ * 각 testcase 마다 설정하는 설정값을 정의합니다.
+ */
 export interface TestCaseConfig {
     pathParams?: Record<string, DSLField<PATH_PARAM_TYPES> | PATH_PARAM_TYPES>
     queryParams?: Record<string, DSLField<QUERY_PARAM_TYPES> | QUERY_PARAM_TYPES>
